Refactor import paths for User and UserInfo models
